Closes #1314
refactor(governance): split governance logic into proposal/voting/execution submodules

Extracted vote_dissolve, propose_contribution_change, and
vote_contribution_change out of the monolithic contract.rs into a new
governance module, following the same thin-facade delegation pattern
already used for payout_executor:

- governance::proposal - creates/stores a pending contribution-amount
  change proposal
- governance::voting - casts dissolve/contribution-change votes and
  checks thresholds
- governance::execution - applies the resulting state change (group
  dissolution + refunds, or the approved contribution amount)

contract.rs's #[contractimpl] functions are now one-line delegators;
the public contract interface and behavior are unchanged. Verified via
cargo check and a release wasm build: 84134 bytes before vs 84183
bytes after (+49 bytes, ~0.06%), confirming contract size is
unaffected.
